refactor server actions, fix spotify logo, setup tailwind
