+++ src/app/actions/reaction.js
@@ -1,8 +1,7 @@
+"use server";
 import { auth } from "@/auth";
-import { getEmbedding } from "./helpers";
 
 export async function postReaction(headline, reaction) {
-  "use server";
   try {
     const session = await auth();
 
@@ -46,3 +45,31 @@ export async function postReaction(headline, reaction) {
     console.log(err);
   }
 }
+
+async function getEmbedding(texts) {
+    try {
+      let body = {
+        texts: texts,
+      };
+      let res = await fetch("https://api.cohere.ai/v1/embed", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+          Accept: "application/json",
+          Authorization: "Bearer " + process.env.COHERE_API_KEY,
+        },
+        body: JSON.stringify(body),
+      });
+  
+      if (!res.ok) {
+        throw new Error(`HTTP error! status: ${res.status}`);
+      }
+  
+      let data = await res.json();
+      return data.embeddings;
+    } catch (error) {
+      console.error("Failed to fetch embeddings:", error);
+      return null;
+    }
+  }
+
