+++ src/app/layout.js
@@ -1,5 +1,5 @@
 import { Inter } from "next/font/google";
-import "@/app/styles/globals.css";
+import "@/app/globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
