+++ src/app/styles/page.css
@@ -14,7 +14,7 @@
   }
   
   .App-header {
-    background:  linear-gradient(180deg, #333, #000);
+    background:  linear-gradient(18deg, #333, #000);
     min-height: 100vh;
     display: flex;
     flex-direction: column;
